pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcap-analyzer"
version = "2.0.0"
description = "Advanced network security analysis tool for PCAP files"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Security",
    "Topic :: System :: Networking :: Monitoring",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "matplotlib>=3.4.0",
    "pyyaml>=6.0",
    "requests>=2.26.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "rich>=12.0.0",  # For better terminal output
    "typer>=0.4.0",  # For better CLI
    "pydantic>=1.9.0",  # For data validation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.1.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.931",
    "pre-commit>=2.17.0",
]
docs = [
    "sphinx>=4.4.0",
    "sphinx-rtd-theme>=1.0.0",
]
viz = [
    "plotly>=5.6.0",
    "dash>=2.3.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/pcap-analyzer"
"Bug Tracker" = "https://github.com/yourusername/pcap-analyzer/issues"

[project.scripts]
pcap-analyzer = "pcap_analysis.cli:main"

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
